rag-system/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Vector DB & Embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Gemini
google-generativeai>=0.3.2

# Data processing
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Logging & Monitoring
loguru>=0.7.2
